Windows 10 detection fix
Added a manifest indicating Windows 10 support to the Windows gPTP build, so that Windows 10 detection will work correctly.  Manifest support is available with CMake 3.4 and later.
Also changed ptp_message.cpp to address a x64 compile warning.
Note that these changes were only tested with Microsoft Visual Studio Premium 2012 Update 5, and not with any other compilers.

+++ daemons/gptp/common/ptp_message.cpp
@@ -1030,7 +1030,7 @@ void PTPMessageFollowUp::processMessage( EtherPort *port )
 
 	/* Adjust local_clock to correspond to sync_arrival */
 	device_sync_time_offset =
-	    TIMESTAMP_TO_NS(device_time) - TIMESTAMP_TO_NS(sync_arrival);
+		(uint32_t) (TIMESTAMP_TO_NS(device_time) - TIMESTAMP_TO_NS(sync_arrival));
 
 	GPTP_LOG_VERBOSE
 	    ("ptp_message::FollowUp::processMessage System time: %u,%u "
